+++ internal/common/metrics/scope_test.go
