+++ internal/common/thrift_util.go
