+++ internal/common/util/stringer_test.go
