+++ internal/common/util/util.go
