+++ internal/internal_decision_state_machine_test.go
@@ -25,7 +25,7 @@ import (
 
 	"github.com/stretchr/testify/require"
 	s "go.uber.org/cadence/.gen/go/shared"
-	"go.uber.org/cadence/common"
+	"go.uber.org/cadence/internal/common"
 )
 
 func Test_TimerStateMachine_CancelBeforeSent(t *testing.T) {
