+++ internal/internal_event_handlers.go
@@ -29,8 +29,8 @@ import (
 
 	"github.com/uber-go/tally"
 	m "go.uber.org/cadence/.gen/go/shared"
-	"go.uber.org/cadence/common"
-	"go.uber.org/cadence/common/metrics"
+	"go.uber.org/cadence/internal/common"
+	"go.uber.org/cadence/internal/common/metrics"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
 )
